Major integrator bug fix, adaptive timescales and other bug fixes and cleanup.
* Fixed a pretty major math error in the integrator.
* Added adaptive timesteps by forcing a reduced timestep for the whole frame
if necessary to maintain a given accuracy.
* Forced a fix time step, even though it's not based on framerate (for testing).
* Particles created are given a slight local time dialation, so some particles don't
quite progress as far as others in a timestep. By setting the start position exactly
the same, this allows observing accumulated error as the particles slowly spread out.
It should create a snaking effect, but obviouly error creaps in and everything eventually
diverges. Blue particles are created with the largest dt, and red are the shortest dt
(so red should lag the blue when all created in the same spot).
* Fixed a bug with polygons not rendering correctly on OS X (this disables the junky anti-aliasing that was on.)
* Cleaned up the code quite a bit. Improved comments on the math. And deleted a lot
of commented out code (since it's now in the git history).