Updated all of the cuda code to doubles from floats.
This was a huge gain in performance since we're bound by a certian
minimum error threshold, as the doubles are more precise!

The current savedTweaks and field lines are setup in a test configuration
that has a dt scale spread over the particles and the particles all
start from the same position and pass right through a wire several times.
This tests that stability of the very close to the wire interactions, as well
as a visual on precision. The fact that we can get a worm like motion of particles
all in a single file line for a decently long time is a good sign that the
numerical integrator is doing an outstanding approximation. It functions really
well when there's not much B-field, but the tight radii caused by the B-fields
generally makes the integrator have to cut down in time significantly.